Clean up ContactInfo and HeadingDetails
